Rename personal portfolio and reposition all portfolios
