+++ src/components/PortfolioManager.tsx
@@ -74,7 +74,7 @@ export const PortfolioManager = () => {
       {/* Portfolio Selection Header */}
       <div className="flex items-center justify-between">
         <div className="flex items-center gap-4">
-          <h2 className="text-2xl font-bold text-foreground">Personal Portfolios</h2>
+          <h2 className="text-2xl font-bold text-foreground">Custom Portfolio</h2>
           <Select
             value={selectedPortfolio?.id || ''}
             onValueChange={(value) => {
